refactor: Class 폴더 구조 변경
