+++ src/pages/Class/Contest/components/index.ts
@@ -1,0 +1,3 @@
+export * from './ContestFormModal';
+export * from './ContestEditModal';
+export * from './ContestProblemCreateModal';
