+++ src/pages/Class/Contest/index.ts
@@ -1,0 +1,3 @@
+export * from './ClassContest';
+export * from './ClassContestProblemList';
+export * from './ClassEditContestList';
