+++ src/pages/Class/Problem/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from './query';
+export * from './useClassProblemListTable';
