+++ src/pages/Class/Problem/index.ts
@@ -8,3 +8,4 @@ export * from './ProblemDescription';
 export * from './ProblemData';
 export * from './ProblemLeaderBoard';
 export * from './ProblemSubmission';
+export * from './ClassProblemList';
