+++ src/pages/Class/StudentManagement/components/index.ts
@@ -1,0 +1,1 @@
+export * from './ClassStudentForm';
