+++ src/pages/Class/StudentManagement/hooks/index.ts
@@ -1,0 +1,3 @@
+export * from './query';
+
+export * from './useStudentForm';
