+++ src/pages/Class/components/index.ts
@@ -1,6 +1,2 @@
 export * from './ClassEditModal';
-export * from './ClassStudentForm';
 export * from './ClassFormModal';
-export * from './ContestFormModal';
-export * from './ContestEditModal';
-export * from './ContestProblemCreateModal';
