+++ src/pages/Class/hooks/index.ts
@@ -3,7 +3,3 @@ export * from './query';
 export * from './useClassListTable';
 export * from './useClassListEditTable';
 export * from './useModalBody';
-export * from './useClassProblemListTable';
-export * from './useStudentForm';
-export * from './useContestProblemListTable';
-export * from './useContestModalBody';
