+++ src/pages/Class/hooks/query/index.ts
@@ -5,22 +5,3 @@ export * from './useClassQuery';
 export * from './useCreateClassMutation';
 export * from './useEditClassMutation';
 export * from './useDeleteClassMutation';
-
-export * from './useProblemListQuery';
-
-export * from './useClassStudentListQuery';
-export * from './useClassTaListQuery';
-export * from './useClassStudentListMutation';
-export * from './useClassTaListMutation';
-
-export * from './useClassContestListQuery';
-export * from './useClassContestProblemQuery';
-
-export * from './useCreateContestMutation';
-export * from './useEditContestVisibleMutation';
-export * from './useDeleteContestMutation';
-export * from './useEditContestMutation';
-
-export * from './useProblemListQuery';
-export * from './useEditProblemPublicMutation';
-export * from './useDeleteProblemMutation';
