+++ src/pages/Class/index.ts
@@ -1,10 +1,7 @@
 export * from './ClassList';
 export * from './ClassListEdit';
 export * from './Class';
-export * from './ClassProblemList';
-export * from './ClassContest';
-export * from './ClassContestProblemList';
-export * from './ClassStudentManagement';
-export * from './ClassEditContestList';
+export * from './Contest';
+export * from './StudentManagement';
 
 export * from './Problem';
